+++ netpyne/utils.py
@@ -10,7 +10,7 @@ from numbers import Number
 from neuron import h
 import importlib
 
-h.load_file("stdrun.hoc") 
+#h.load_file("stdrun.hoc") 
 
 def getSecName (sec, dirCellSecNames = None):
     if dirCellSecNames is None: dirCellSecNames = {}
@@ -205,7 +205,7 @@ def importCell (fileName, cellName, cellArgs = None, cellInstance = False):
 
 def importCellsFromNet (netParams, fileName, labelList, condsList, cellNamesList, importSynMechs):
     h.initnrn()
-
+    
     ''' Import cell from HOC template or python file into framework format (dict of sections, with geom, topol, mechs, syns)'''
     if fileName.endswith('.hoc') or fileName.endswith('.tem'):
         print 'Importing from .hoc network not yet supported'
@@ -239,14 +239,16 @@ def importCellsFromNet (netParams, fileName, labelList, condsList, cellNamesList
         print '\nImporting %s from %s ...'%(cellName, fileName)
         exec('cell = tempModule' + '.' + cellName)
         #cell = getattr(modulePointer, cellName) # get cell object
-        secs, secLists, synMechs = getCellParams(cell)
+        varList = mechVarList()
+        origGlob = getGlobals(varList['mechs'].keys()+varList['pointps'].keys())
+        secs, secLists, synMechs = getCellParams(cell, varList, origGlob)
         cellRule = {'conds': conds, 'secs': secs, 'secLists': secLists}
         netParams.addCellParams(label, cellRule)
         if importSynMechs:
             for synMech in synMechs: netParams.addSynMechParams(synMech.pop('label'), synMech)
 
 
-def getCellParams(cell, varList, origGlob):
+def getCellParams(cell, varList={}, origGlob={}):
     dirCell = dir(cell)
 
     if 'all_sec' in dirCell: